chore: centralize all protobuf formatting methods in StorageV2ProtoUtils
Update GrpcPlainRequestLoggingInterceptor to access StorageV2ProtoUtils.fmtProto via PackagePrivateMethodWorkarounds
